# jws: improve performance for SplitCompact/SplitCompactString