name            old time/op    new time/op    delta
SplitCompat-16    76.5ns ± 5%    21.1ns ± 1%   -72.40%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
SplitCompat-16      272B ± 0%        0B       -100.00%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
SplitCompat-16      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)