@drakkan 2025-04-19
Thanks! Could this be applied to v2 as well, please?