@lestrrat 2025-04-20
@drakkan I can't cherry-pick this change to v2 due to some problems (*), can you please create a PR against v2? To be clear, they are minor problems that I can deal myself, but I thought you might want to make the commits so that your name will show up in the records, since you were one who came up with this.

Also, I'd gratefully accept it if you submitted a PR for JWE.

(*) biggest one is that there's no jws.ParseError in v2